fixexes for Visual Studio and /W4 switch

+++ JSON/src/pdjson.c
@@ -123,7 +123,7 @@ static void init(json_stream *json)
     json->next = 0;
 
     json->stack = NULL;
-    json->stack_top = -1;
+    json->stack_top = (size_t)(-1);
     json->stack_size = 0;
 
     json->data.string = NULL;
@@ -751,7 +751,7 @@ enum json_type json_next(json_stream *json)
 
 void json_reset(json_stream *json)
 {
-    json->stack_top = -1;
+	json->stack_top = (size_t)(-1);
     json->ntokens = 0;
     json->flags &= ~JSON_FLAG_ERROR;
     json->errmsg[0] = '\0';
